pyproject.toml
[project]
name = "pipeline-for-prompthallucinationreductions"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "requests",
    "python-dotenv",
]
